backend/requirements.txt
# FastAPI and web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0

# LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.20
langsmith>=0.0.69

# Google AI
google-generativeai>=0.3.2

# Database and ORM
sqlalchemy>=2.0.23
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Redis and Celery
redis>=5.0.1
celery>=5.3.4

# Data processing
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.3.2
xgboost>=2.0.2
lightgbm>=4.1.0

# Visualization
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.2.1

# HTTP requests
httpx>=0.25.2
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Monitoring and logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Jupyter notebook generation
nbformat>=5.9.2
jupyter>=1.0.0

# Report generation
jinja2>=3.1.2
markdown>=3.5.1

# Web scraping for data discovery
beautifulsoup4>=4.12.2
requests-html>=0.10.0

# Additional ML libraries
imbalanced-learn>=0.11.0
optuna>=3.4.0
shap>=0.43.0